Keep file tasks with ':' at top level
With #232 you can now create tasks with namespaces from the task name
alone.  This did not take into account that file tasks can contain a ':'
which meant file tasks would not exist at top-level.  This was likely
missed due to a test that only ran on windows.

Now the namespace check is only performed for non-file tasks.

Fixes #262