Fix infinite loop when cleaning files on windows
When checking if nonexistent file was deleted on windows the
tree-walking loop did not terminate properly.  Instead of checking for
no change for the path it looked for "." or "/", the latter of which
will not match a windows drive letter.

Now if the previous path is seen again the loop stops.