Ignore falsey dependencies
When creating Rake tasks programatically, it could so happen that the
prerequisites reach the point of creation as an argument. This change
teaches the Rake DSL to ignore falsey (`nil' or `false') values when
they are specified as the dependencies; this way, programmatic creation
does not have to deal with a special case.