Prevent useless relinking when packaging
Previously rake would relink everything in the package directory even
when no file had changed.  This occurred because rake was told the
package directory was a file so it would attempt to recreate it by
linking all the files.

Now Rake knows the package directory is a directory and will not attempt
to recreated it unless the package file list changes.

Fixes #276