[WIP] Synchronized stream output while invoking concurrent prerequisites
NOTE: This doesn't fully solve the problem.  See below.

This implements the IO stream wrapper idea I had.  It's in synchronized_output.rb.  For now,
the wrapping is done in Task#invoke_prerequisites_concurrently, though there's commented-out
code that can do it in Application#run instead.  I'm not positive which way is best; thoughts
appreciated.

The code and tests still need a bit of love, but given that this approach doesn't work, I haven't yet
put the effort in to finish cleaning it up.

The problem with this implementation is that not all output to stdout and stderr is done at the Ruby
level and so it doesn't go through the wrapper class.  For example in MRI, Kernel::print and Kernel::puts
call down into C before forwarding on to stdout, so only the lower-level `write` calls go through the
wrapper.  Thus, the `print` or `puts` calls are not locked for their entire duration.  Another issue occurs
when using `sh` to call an external program that writes to stdout or stderr; these writes also would not
be locked properly.

I believe that a different approach is required to solve this problem, but I don't have a good idea what
that might be.  I'd love to hear some ideas, and I'm willing to do the legwork on it.