Correctly handle bad encoding in exception messages
TraceOutput#trace_on has a bug that causes the matching to fail if a
string is incorrectly encoded, for example encoded as 'US-ASCII' but
contains unicode characters. This causes a new exception to be thrown,
losing the stack trace of the previous exception.
This can occur, for example, in Ruby 1.9 when one of the dependencies of
a project includes a file that contains unicode but does not specify an
encoding (as the default for Ruby 1.9 is US-ASCII). This was the case in
a certain version of the Puma web server, which was fixed in
https://github.com/puma/puma/pull/924 but was very difficult to debug due
to the incorrect stack trace provided by rake.