Allow implicit ARGV for Rake::Application#run
This allows users to pass in an explicit ARGV for rake instead of
changing their application's ARGV.

While the previous commit allows users to create a Rake application
explicitly by calling parts of Rake::Application#init directly, some
users will not want to dig through the source to set up a Rake
application.

Swapping ARGV is confusing and error-prone, so this allows them to
specify a set of Rake options directly.