Lazily load `set`
This file seems required by `rake` by default. By lazily loading `set`,
usages of `rake` that don't use this part of `rake`, don't unnecessarily
activate the `set` gem.

This seems really minor, I know. But when testing rubygems & bundler,
it's interesting for us to be in full control of the gems that are
loaded, to make sure that we don't unintentionally activate gems causing
our users to be unable to select the version of those gems that they
need.