Redirect from old location
The previous commit moved a page, so here we add a redirect at the old
location pointing to the new. There appears to be no pre-existing
mechanism to accommodate redirects, so this adds a step to the build
process that copies the raw HTML representing the redirecting page into
the target directory.